requirements.txt
asgiref==3.5.2
coverage==6.4.1
Django==4.0.4
django-debug-toolbar==3.5.0
Pillow==9.1.1
psycopg2==2.9.3
sorl-thumbnail==12.8.0
sqlparse==0.4.2
